requirements.txt

matplotlib.pyplot
numpy
pandas
scipy.signal
seaborn
networkx
pytest
